requirements.txt
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
et_xmlfile==2.0.0
idna==3.11
lxml==6.0.2
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.0
pillow==12.1.0
python-dateutil==2.9.0.post0
requests==2.32.5
screeninfo==0.8.1
six==1.17.0
soupsieve==2.8.3
tkinterdnd2==0.4.3
typing_extensions==4.15.0
urllib3==2.6.3
xlsxwriter==3.2.9
